Remove direct vector example, replace with interrupt latency test
