+++ hil-test/Cargo.toml
@@ -37,6 +37,10 @@ name    = "gpio"
 harness = false
 
 [[test]]
+name    = "interrupt"
+harness = false
+
+[[test]]
 name    = "spi_full_duplex"
 harness = false
